layering fixes, add grid to main view, reorganize info-panel
